+++ SW2/projects/FORTH_c_003/src/tools/sd_getc.h
@@ -1,0 +1,13 @@
+/* vim: ft=cpp showbreak=--»\  noexpandtab fileencoding=utf-8 nomodified wrap textwidth=0 foldmethod=marker foldmarker={{{,}}} foldcolumn=4 ruler showcmd lcs=tab\:|- list: tabstop=8 linebreak  
+ * */
+// ,,g = gcc, exactly one space after "set"
+//
+#pragma once
+
+
+#include "ff.h"
+#include "getc.h"
+#include <stddef.h>
+
+uint8_t sd_getc(void *ptr, char *c);	// exported, source of chars
+void C_SD_LOAD(char * C_filename);	// word SD.LOAD, call with C string in RAM